simcom without rtos

+++ v_NATIVE/Src/delays.c
@@ -1,10 +1,13 @@
 #include "delays.h"
 #include <stdint.h>
+#include "timer16.h"
 #include "timer6.h"
-void usDelay(uint32_t Delay){
+#include "core.h"
+
+// warnig appropriate queues needs to be initialized
+
+void usDelay(uint64_t Delay){
  usDelayTime = 0; 
  while(usDelayTime < Delay){
-  
  }
-
 }
